Move diplomacy-related files to subfolder
