+++ Vic2ToHoI4/Source/HOI4World/Diplomacy/HoI4Agreement.h
@@ -26,7 +26,7 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.*/
 
 
 #include "Date.h"
-#include "../V2World/Agreement.h"
+#include "../../V2World/Agreement.h"
 #include <string>
 using namespace std;
