+++ Vic2ToHoI4/Source/HOI4World/Diplomacy/HoI4Diplomacy.cpp
@@ -24,7 +24,7 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.*/
 #include "HoI4Diplomacy.h"
 #include "HoI4Agreement.h"
 #include "Log.h"
-#include "../Configuration.h"
+#include "../../Configuration.h"
 #include "OSCompatibilityLayer.h"
 #include <fstream>
