+++ Vic2ToHoI4/Source/HOI4World/Diplomacy/HoI4WarCreator.cpp
@@ -22,16 +22,16 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.*/
 
 
 #include "HoI4WarCreator.h"
-#include "Events/Events.h"
 #include "HoI4Faction.h"
-#include "HoI4Focus.h"
-#include "HoI4World.h"
-#include "HoI4Localisation.h"
-#include "Map/MapData.h"
-#include "../Mappers/ProvinceDefinitions.h"
-#include "../V2World/Country.h"
-#include "../V2World/Party.h"
-#include "../V2World/World.h"
+#include "../Events/Events.h"
+#include "../HoI4Focus.h"
+#include "../HoI4Localisation.h"
+#include "../HoI4World.h"
+#include "../Map/MapData.h"
+#include "../../Mappers/ProvinceDefinitions.h"
+#include "../../V2World/Country.h"
+#include "../../V2World/Party.h"
+#include "../../V2World/World.h"
 #include "Log.h"
