+++ Vic2ToHoI4/Source/HOI4World/HoI4Country.cpp
@@ -1,8 +1,8 @@
+#include "Diplomacy/HoI4War.h"
 #include "HoI4Country.h"
 #include "HoI4World.h"
 #include "Log.h"
 #include "HoI4Localisation.h"
-#include "HoI4War.h"
 #include "Leaders/Advisor.h"
 #include "Names.h"
 #include "MilitaryMappings/MilitaryMappings.h"
