+++ Vic2ToHoI4/Source/HOI4World/HoI4World.cpp
@@ -1,15 +1,16 @@
+#include "Diplomacy/HoI4Agreement.h"
+#include "Diplomacy/HoI4Diplomacy.h"
+#include "Diplomacy/HoI4Faction.h"
+#include "Diplomacy/HoI4WarCreator.h"
 #include "HoI4World.h"
 #include "Log.h"
 #include "OSCompatibilityLayer.h"
 #include "../Configuration.h"
 #include "../V2World/Diplomacy.h"
 #include "../V2World/Party.h"
-#include "HoI4Agreement.h"
 #include "HoI4Country.h"
 #include "Decisions/Decisions.h"
-#include "HoI4Diplomacy.h"
 #include "Events/Events.h"
-#include "HoI4Faction.h"
 #include "HoI4FocusTree.h"
 #include "HOI4Ideology.h"
 #include "HoI4Localisation.h"
@@ -18,7 +19,6 @@
 #include "Leaders/Advisor.h"
 #include "Leaders/IdeologicalAdvisors.h"
 #include "Names.h"
-#include "HoI4WarCreator.h"
 #include "Map/Buildings.h"
 #include "Map/StrategicRegion.h"
 #include "Map/SupplyZones.h"
