+++ Vic2ToHoI4/Source/Hoi4Outputter/Hoi4CountryOutputter.cpp
@@ -4,9 +4,9 @@
 #include "Navies/OutLegacyNavyNames.h"
 #include "Navies/OutMtgNavyNames.h"
 #include "ShipTypes/ShipVariantsOutputter.h"
+#include "../HOI4World/Diplomacy/HoI4Faction.h"
 #include "../HOI4World/DivisionTemplate.h"
 #include "../HOI4World/HoI4Country.h"
-#include "../HOI4World/HoI4Faction.h"
 #include "../HOI4World/HoI4FocusTree.h"
 #include "../HOI4World/Names.h"
 #include "../HOI4World/Navies/NavyNames.h"
