+++ Vic2ToHoI4/Vic2ToHoI4.vcxproj
@@ -224,8 +224,6 @@
     <ClCompile Include="Source\Hoi4Outputter\ShipTypes\ShipVariantsOutputter.cpp" />
     <ClCompile Include="Source\Hoi4Outputter\States\HoI4StateOutputter.cpp" />
     <ClCompile Include="Source\Hoi4Outputter\States\HoI4StatesOutputter.cpp" />
-    <ClCompile Include="Source\HOI4World\AIPeace.cpp" />
-    <ClCompile Include="Source\HOI4World\AIPeaces.cpp" />
     <ClCompile Include="Source\HOI4World\Decisions\Decision.cpp" />
     <ClCompile Include="Source\HOI4World\Decisions\Decisions.cpp" />
     <ClCompile Include="Source\HOI4World\Decisions\DecisionsCategory.cpp" />
@@ -237,24 +235,26 @@
     <ClCompile Include="Source\HOI4World\Decisions\NavalTreatyDecisions.cpp" />
     <ClCompile Include="Source\HOI4World\Decisions\PoliticalDecisions.cpp" />
     <ClCompile Include="Source\HOI4World\Decisions\StabilityWarSupportDecisions.cpp" />
+    <ClCompile Include="Source\HOI4World\Diplomacy\AIPeace.cpp" />
+    <ClCompile Include="Source\HOI4World\Diplomacy\AIPeaces.cpp" />
+    <ClCompile Include="Source\HOI4World\Diplomacy\HoI4Agreement.cpp" />
+    <ClCompile Include="Source\HOI4World\Diplomacy\HoI4Diplomacy.cpp" />
+    <ClCompile Include="Source\HOI4World\Diplomacy\HoI4Relations.cpp" />
+    <ClCompile Include="Source\HOI4World\Diplomacy\HoI4War.cpp" />
+    <ClCompile Include="Source\HOI4World\Diplomacy\HoI4WarCreator.cpp" />
     <ClCompile Include="Source\HOI4World\Division.cpp" />
     <ClCompile Include="Source\HOI4World\DivisionTemplate.cpp" />
     <ClCompile Include="Source\HOI4World\Events\Event.cpp" />
     <ClCompile Include="Source\HOI4World\Events\EventOption.cpp" />
     <ClCompile Include="Source\HOI4World\Events\Events.cpp" />
-    <ClCompile Include="Source\HOI4World\HoI4Agreement.cpp" />
     <ClCompile Include="Source\HOI4World\HoI4Airforce.cpp" />
     <ClCompile Include="Source\HOI4World\HoI4Army.cpp" />
     <ClCompile Include="Source\HOI4World\HoI4Country.cpp" />
-    <ClCompile Include="Source\HOI4World\HoI4Diplomacy.cpp" />
     <ClCompile Include="Source\HOI4World\HoI4Focus.cpp" />
     <ClCompile Include="Source\HOI4World\HoI4FocusTree.cpp" />
     <ClCompile Include="Source\HOI4World\HOI4Ideology.cpp" />
     <ClCompile Include="Source\HOI4World\HoI4Localisation.cpp" />
-    <ClCompile Include="Source\HOI4World\HoI4Relations.cpp" />
     <ClCompile Include="Source\HOI4World\HOI4Version.cpp" />
-    <ClCompile Include="Source\HOI4World\HoI4War.cpp" />
-    <ClCompile Include="Source\HOI4World\HoI4WarCreator.cpp" />
     <ClCompile Include="Source\HOI4World\HoI4World.cpp" />
     <ClCompile Include="Source\HOI4World\Idea.cpp" />
     <ClCompile Include="Source\HOI4World\IdeaGroup.cpp" />
@@ -394,8 +394,6 @@
     <ClInclude Include="Source\Hoi4Outputter\ShipTypes\ShipVariantsOutputter.h" />
     <ClInclude Include="Source\Hoi4Outputter\States\HoI4StateOutputter.h" />
     <ClInclude Include="Source\Hoi4Outputter\States\HoI4StatesOutputter.h" />
-    <ClInclude Include="Source\HOI4World\AIPeace.h" />
-    <ClInclude Include="Source\HOI4World\AIPeaces.h" />
     <ClInclude Include="Source\HOI4World\Decisions\Decision.h" />
     <ClInclude Include="Source\HOI4World\Decisions\Decisions.h" />
     <ClInclude Include="Source\HOI4World\Decisions\DecisionsCategory.h" />
@@ -407,25 +405,27 @@
     <ClInclude Include="Source\HOI4World\Decisions\NavalTreatyDecisions.h" />
     <ClInclude Include="Source\HOI4World\Decisions\PoliticalDecisions.h" />
     <ClInclude Include="Source\HOI4World\Decisions\StabilityWarSupportDecisions.h" />
+    <ClInclude Include="Source\HOI4World\Diplomacy\AIPeace.h" />
+    <ClInclude Include="Source\HOI4World\Diplomacy\AIPeaces.h" />
+    <ClInclude Include="Source\HOI4World\Diplomacy\HoI4Agreement.h" />
+    <ClInclude Include="Source\HOI4World\Diplomacy\HoI4Diplomacy.h" />
+    <ClInclude Include="Source\HOI4World\Diplomacy\HoI4Faction.h" />
+    <ClInclude Include="Source\HOI4World\Diplomacy\HoI4Relations.h" />
+    <ClInclude Include="Source\HOI4World\Diplomacy\HoI4War.h" />
+    <ClInclude Include="Source\HOI4World\Diplomacy\HoI4WarCreator.h" />
     <ClInclude Include="Source\HOI4World\Division.h" />
     <ClInclude Include="Source\HOI4World\DivisionTemplate.h" />
     <ClInclude Include="Source\HOI4World\Events\Event.h" />
     <ClInclude Include="Source\HOI4World\Events\EventOption.h" />
     <ClInclude Include="Source\HOI4World\Events\Events.h" />
-    <ClInclude Include="Source\HOI4World\HoI4Agreement.h" />
     <ClInclude Include="Source\HOI4World\HoI4Airforce.h" />
     <ClInclude Include="Source\HOI4World\HoI4Army.h" />
     <ClInclude Include="Source\HOI4World\HoI4Country.h" />
-    <ClInclude Include="Source\HOI4World\HoI4Diplomacy.h" />
-    <ClInclude Include="Source\HOI4World\HoI4Faction.h" />
     <ClInclude Include="Source\HOI4World\HoI4Focus.h" />
     <ClInclude Include="Source\HOI4World\HoI4FocusTree.h" />
     <ClInclude Include="Source\HOI4World\HOI4Ideology.h" />
     <ClInclude Include="Source\HOI4World\HoI4Localisation.h" />
-    <ClInclude Include="Source\HOI4World\HoI4Relations.h" />
     <ClInclude Include="Source\HOI4World\HOI4Version.h" />
-    <ClInclude Include="Source\HOI4World\HoI4War.h" />
-    <ClInclude Include="Source\HOI4World\HoI4WarCreator.h" />
     <ClInclude Include="Source\HOI4World\HoI4World.h" />
     <ClInclude Include="Source\HOI4World\Idea.h" />
     <ClInclude Include="Source\HOI4World\IdeaGroup.h" />
