+++ Vic2ToHoI4/Vic2ToHoI4.vcxproj.filters
@@ -94,6 +94,9 @@
     <Filter Include="Hoi4Outputter\ScriptedTriggers">
       <UniqueIdentifier>{dcc00db9-91ca-4d39-816a-75b884ceb428}</UniqueIdentifier>
     </Filter>
+    <Filter Include="HoI4World\Diplomacy">
+      <UniqueIdentifier>{380fc73c-dfd8-40bc-b0f0-ea0b3d25f1d6}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\common_items\Date.cpp">
@@ -108,13 +111,7 @@
     <ClCompile Include="Source\HOI4World\HoI4Country.cpp">
       <Filter>HoI4World</Filter>
     </ClCompile>
-    <ClCompile Include="Source\HOI4World\HoI4Diplomacy.cpp">
-      <Filter>HoI4World</Filter>
-    </ClCompile>
     <ClCompile Include="Source\HOI4World\HoI4Localisation.cpp">
-      <Filter>HoI4World</Filter>
-    </ClCompile>
-    <ClCompile Include="Source\HOI4World\HoI4Relations.cpp">
       <Filter>HoI4World</Filter>
     </ClCompile>
     <ClCompile Include="Source\HOI4World\HoI4World.cpp">
@@ -152,17 +149,11 @@
     <ClCompile Include="Source\HOI4World\HoI4Focus.cpp">
       <Filter>HoI4World</Filter>
     </ClCompile>
-    <ClCompile Include="Source\HOI4World\HoI4WarCreator.cpp">
-      <Filter>HoI4World</Filter>
-    </ClCompile>
     <ClCompile Include="..\common_items\CardinalToOrdinal.cpp">
       <Filter>CommonItems</Filter>
     </ClCompile>
     <ClCompile Include="Source\Mappers\ProvinceDefinitions.cpp">
       <Filter>Mappers</Filter>
-    </ClCompile>
-    <ClCompile Include="Source\HOI4World\HoI4Agreement.cpp">
-      <Filter>HoI4World</Filter>
     </ClCompile>
     <ClCompile Include="Source\Mappers\GraphicsMapper.cpp">
       <Filter>Mappers</Filter>
@@ -228,12 +219,6 @@
       <Filter>HoI4World</Filter>
     </ClCompile>
     <ClCompile Include="Source\HOI4World\IdeaGroup.cpp">
-      <Filter>HoI4World</Filter>
-    </ClCompile>
-    <ClCompile Include="Source\HOI4World\AIPeaces.cpp">
-      <Filter>HoI4World</Filter>
-    </ClCompile>
-    <ClCompile Include="Source\HOI4World\AIPeace.cpp">
       <Filter>HoI4World</Filter>
     </ClCompile>
     <ClCompile Include="Source\HOI4World\OnActions.cpp">
@@ -338,9 +323,6 @@
     </ClCompile>
     <ClCompile Include="Source\V2World\Wars\WarGoal.cpp">
       <Filter>Vic2World\Wars</Filter>
-    </ClCompile>
-    <ClCompile Include="Source\HOI4World\HoI4War.cpp">
-      <Filter>HoI4World</Filter>
     </ClCompile>
     <ClCompile Include="Source\HOI4World\States\HoI4State.cpp">
       <Filter>HoI4World\States</Filter>
@@ -582,6 +564,27 @@
     <ClCompile Include="Source\HOI4World\ScriptedTriggers\ScriptedTriggersUpdater.cpp">
       <Filter>HoI4World\ScriptedTriggers</Filter>
     </ClCompile>
+    <ClCompile Include="Source\HOI4World\Diplomacy\AIPeace.cpp">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClCompile>
+    <ClCompile Include="Source\HOI4World\Diplomacy\AIPeaces.cpp">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClCompile>
+    <ClCompile Include="Source\HOI4World\Diplomacy\HoI4Agreement.cpp">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClCompile>
+    <ClCompile Include="Source\HOI4World\Diplomacy\HoI4Diplomacy.cpp">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClCompile>
+    <ClCompile Include="Source\HOI4World\Diplomacy\HoI4Relations.cpp">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClCompile>
+    <ClCompile Include="Source\HOI4World\Diplomacy\HoI4War.cpp">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClCompile>
+    <ClCompile Include="Source\HOI4World\Diplomacy\HoI4WarCreator.cpp">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClCompile>
     <ClCompile Include="Source\Hoi4Outputter\Navies\OutLegacyNavyNames.cpp">
       <Filter>Hoi4Outputter\Navies</Filter>
     </ClCompile>
@@ -608,13 +611,7 @@
     <ClInclude Include="Source\HOI4World\HoI4Country.h">
       <Filter>HoI4World</Filter>
     </ClInclude>
-    <ClInclude Include="Source\HOI4World\HoI4Diplomacy.h">
-      <Filter>HoI4World</Filter>
-    </ClInclude>
     <ClInclude Include="Source\HOI4World\HoI4Localisation.h">
-      <Filter>HoI4World</Filter>
-    </ClInclude>
-    <ClInclude Include="Source\HOI4World\HoI4Relations.h">
       <Filter>HoI4World</Filter>
     </ClInclude>
     <ClInclude Include="Source\HOI4World\HoI4World.h">
@@ -642,16 +639,10 @@
     <ClInclude Include="Source\HOI4World\HOI4Version.h">
       <Filter>HoI4World</Filter>
     </ClInclude>
-    <ClInclude Include="Source\HOI4World\HoI4Faction.h">
-      <Filter>HoI4World</Filter>
-    </ClInclude>
     <ClInclude Include="Source\HOI4World\HoI4FocusTree.h">
       <Filter>HoI4World</Filter>
     </ClInclude>
     <ClInclude Include="Source\HOI4World\HoI4Focus.h">
-      <Filter>HoI4World</Filter>
-    </ClInclude>
-    <ClInclude Include="Source\HOI4World\HoI4WarCreator.h">
       <Filter>HoI4World</Filter>
     </ClInclude>
     <ClInclude Include="..\common_items\CardinalToOrdinal.h">
@@ -659,9 +650,6 @@
     </ClInclude>
     <ClInclude Include="Source\Mappers\ProvinceDefinitions.h">
       <Filter>Mappers</Filter>
-    </ClInclude>
-    <ClInclude Include="Source\HOI4World\HoI4Agreement.h">
-      <Filter>HoI4World</Filter>
     </ClInclude>
     <ClInclude Include="Source\Mappers\GraphicsMapper.h">
       <Filter>Mappers</Filter>
@@ -727,12 +715,6 @@
       <Filter>HoI4World</Filter>
     </ClInclude>
     <ClInclude Include="Source\HOI4World\IdeaGroup.h">
-      <Filter>HoI4World</Filter>
-    </ClInclude>
-    <ClInclude Include="Source\HOI4World\AIPeaces.h">
-      <Filter>HoI4World</Filter>
-    </ClInclude>
-    <ClInclude Include="Source\HOI4World\AIPeace.h">
       <Filter>HoI4World</Filter>
     </ClInclude>
     <ClInclude Include="Source\HOI4World\OnActions.h">
@@ -858,9 +840,6 @@
     </ClInclude>
     <ClInclude Include="Source\V2World\Wars\WarGoal.h">
       <Filter>Vic2World\Wars</Filter>
-    </ClInclude>
-    <ClInclude Include="Source\HOI4World\HoI4War.h">
-      <Filter>HoI4World</Filter>
     </ClInclude>
     <ClInclude Include="Source\HOI4World\States\HoI4State.h">
       <Filter>HoI4World\States</Filter>
@@ -1107,6 +1086,30 @@
     </ClInclude>
     <ClInclude Include="Source\HOI4World\ScriptedTriggers\ScriptedTriggersUpdater.h">
       <Filter>HoI4World\ScriptedTriggers</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\HOI4World\Diplomacy\AIPeace.h">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\HOI4World\Diplomacy\AIPeaces.h">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\HOI4World\Diplomacy\HoI4Agreement.h">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\HOI4World\Diplomacy\HoI4Diplomacy.h">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\HOI4World\Diplomacy\HoI4Relations.h">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\HOI4World\Diplomacy\HoI4War.h">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\HOI4World\Diplomacy\HoI4WarCreator.h">
+      <Filter>HoI4World\Diplomacy</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\HOI4World\Diplomacy\HoI4Faction.h">
+      <Filter>HoI4World\Diplomacy</Filter>
     </ClInclude>
     <ClInclude Include="Source\Hoi4Outputter\Navies\OutLegacyNavyNames.h">
       <Filter>Hoi4Outputter\Navies</Filter>
