+++ Vic2ToHoI4Tests/Vic2ToHoI4Tests-Google.vcxproj
@@ -89,8 +89,6 @@
     <ClCompile Include="..\Vic2ToHoI4\Source\Hoi4Outputter\ShipTypes\ShipVariantsOutputter.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\Hoi4Outputter\States\HoI4StateOutputter.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\Hoi4Outputter\States\HoI4StatesOutputter.cpp" />
-    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\AIPeace.cpp" />
-    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\AIPeaces.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Decisions\Decision.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Decisions\Decisions.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Decisions\DecisionsCategory.cpp" />
@@ -102,24 +100,26 @@
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Decisions\NavalTreatyDecisions.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Decisions\PoliticalDecisions.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Decisions\StabilityWarSupportDecisions.cpp" />
+    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Diplomacy\AIPeace.cpp" />
+    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Diplomacy\AIPeaces.cpp" />
+    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Diplomacy\HoI4Agreement.cpp" />
+    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Diplomacy\HoI4Diplomacy.cpp" />
+    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Diplomacy\HoI4Relations.cpp" />
+    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Diplomacy\HoI4War.cpp" />
+    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Diplomacy\HoI4WarCreator.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Division.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\DivisionTemplate.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Events\Event.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Events\EventOption.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Events\Events.cpp" />
-    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HoI4Agreement.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HoI4Airforce.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HoI4Army.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HoI4Country.cpp" />
-    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HoI4Diplomacy.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HoI4Focus.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HoI4FocusTree.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HOI4Ideology.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HoI4Localisation.cpp" />
-    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HoI4Relations.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HOI4Version.cpp" />
-    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HoI4War.cpp" />
-    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HoI4WarCreator.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\HoI4World.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Idea.cpp" />
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\IdeaGroup.cpp" />
